:recycle: move helpers functions near their applications
